refactor: Update group store to use int64 for IDs
Updated the GroupStore to use int64 for the group ID in its methods. This change ensures consistency with the updated models and database schema, preventing potential data type mismatch errors.

+++ backend/internal/store/group_store.go
@@ -33,12 +33,12 @@ func (s *groupStore) CreateGroup(group *models.Group) (*models.Group, error) {
 		return nil, err
 	}
 
-	group.ID = int(id)
+	group.ID = id
 
 	return group, nil
 }
 
-func (s *groupStore) GetGroupByID(groupID int) (*models.Group, error) {
+func (s *groupStore) GetGroupByID(groupID int64) (*models.Group, error) {
 	var group models.Group
 	err := s.db.QueryRow("SELECT id, creator_id, title, description, privacy, created_at FROM groups WHERE id = ?", groupID).Scan(
 		&group.ID,
